feat: 支持 MySQL，PostgreSQL。添加 Docker 部署教程

+++ internal/types/shortener.go
@@ -6,7 +6,7 @@ type Url struct {
 	ShortURL    string `json:"short_url"`
 	OriginalURL string `json:"original_url"`
 	Describe    string `json:"describe"`
-	Status      int32  `json:"status"`
+	Status      int8   `json:"status"`
 	CreatedAt   int64  `json:"created_at"`
 	UpdatedAt   int64  `json:"updated_at"`
 }
